CategoryController 4 new Methods
Create the CategoryController methods:

getAllCategories
saveCategory
updateCategory
deleteCategory

Also i have created the CategoryService method for update named... updateCategory.

And final change, all methods now have a Return to do getAll when finish the action.

This getAllCategories do a pull in the database to show a List of Categories in Alphabetical order.